backend/requirements.txt
alembic==1.12.1
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.3.0
cffi==1.17.1
click==8.2.1
cryptography==45.0.5
dnspython==2.7.0
ecdsa==0.19.1
elastic-transport==8.15.1
elasticsearch==8.11.1
email_validator==2.2.0
fastapi==0.104.1
greenlet==3.2.3
h11>=0.13.0,<0.15.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
psycopg2-binary==2.9.9
python-dotenv==1.1.1
python-jose==3.3.0
python-multipart==0.0.6
redis==4.6.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.14.1
uvicorn==0.24.0
# Test dependencies
pytest>=7.0.0
pytest-asyncio>=0.18.0
httpx>=0.23.0,<0.25.0
pytest-cov>=4.0.0
faker>=19.0.0 